refactor: organize modules into packages
