+++ README_OPERATOR.md
@@ -531,7 +531,7 @@ python scripts/show_emotion_glyph.py
 
 ### Voice Aura FX
 
-`voice_aura.py` selects a reverb and delay preset for the active emotion and
+`audio/voice_aura.py` selects a reverb and delay preset for the active emotion and
 applies it with `sox` when available. If `sox` is missing the script falls back
 to `pydub`.  Optional RAVE or NSynth checkpoints can further blend the timbre
 via `dsp_engine.rave_morph` or `nsynth_interpolate`.
