+++ src/core/__init__.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 from .contracts import EmotionAnalyzerService, MemoryLoggerService
 from .emotion_analyzer import EmotionAnalyzer
+from .feedback_logging import append_feedback, load_feedback
 from .memory_logger import MemoryLogger
 
 __all__ = [
@@ -11,4 +12,6 @@ __all__ = [
     "MemoryLogger",
     "EmotionAnalyzerService",
     "MemoryLoggerService",
+    "append_feedback",
+    "load_feedback",
 ]
