+++ src/dashboard/system_monitor.py
